Conditionalize migration for empty databases on #1992

+++ db/migrate/20200819161133_clean_duplicate_collection_owners.rb
@@ -1,7 +1,9 @@
 class CleanDuplicateCollectionOwners < ActiveRecord::Migration[6.0]
   def change
-    owner_map = CollectionOwner.distinct.all.to_a
-    CollectionOwner.delete_all
-    CollectionOwner.insert_all!(owner_map.map{|obj| obj.attributes})
+    owner_map = CollectionOwner.select('DISTINCT user_id, collection_id').to_a
+    if owner_map.size > 0
+      CollectionOwner.delete_all
+      CollectionOwner.insert_all!(owner_map.map{|obj| obj.attributes})
+    end
   end
 end
